third_party/WebKit/LayoutTests/platform/win/virtual/threaded/fast/scroll-behavior/smooth-scroll/mousewheel-scroll-interrupted-expected.txt

This test ensures that input driven user scrolls get cancelled by non-input driven scrolls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getScrollAnimationState(document) is RUNNING_ON_COMPOSITOR
PASS isOrWillBeCancelled(document) is true
FAIL document.scrollingElement.scrollTop < 40 && document.scrollingElement.scrollLeft == 0 failed to change to true in 0.5 seconds.
PASS successfullyParsed is true

TEST COMPLETE
